+++ R/options.R
@@ -1,0 +1,1 @@
+unique_globals <- function() getOption("roxyglobals.unique") %??% FALSE